Forward-renderer refactoring - moved shadow related code out
